Update installation script
- Update README.md
- Ubuntu
  - Set default user for wsl
- Windows
  - Install Windows Terminal
  - Install fonts for terminal
  - Change default font for Windows Terminal
  - Limit wsl's ram usage
  - Add reboot prompt when executing wsl_presiquite.ps1

+++ windows_presiquite.ps1
@@ -8,4 +8,12 @@ dism.exe /online /enable-feature /featurename:Microsoft-Windows-Subsystem-Linux 
 # Enable virtual machine platform
 dism.exe /online /enable-feature /featurename:VirtualMachinePlatform /all /norestart
 
-Restart-Computer -Confirm:$true
+$Prompt = Read-Host -Prompt "Do you want to reboot now? [Y/n] "
+
+if (($Prompt -eq "Y") -or ($Prompt -eq "y") -or ($Prompt -eq "")) {
+    Restart-Computer -Confirm:$true
+}
+else {
+    Write-Host "You have to manually reboot to execute remainders!" `
+        -ForegroundColor Red -BackgroundColor Gray
+}
